+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationVisitor.java
@@ -34,7 +34,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Set;
 import java.util.StringJoiner;
 
 import javax.lang.model.element.AnnotationMirror;
@@ -253,7 +252,7 @@ public class InitializationVisitor extends BaseTypeVisitor<InitializationAnnotat
         commonAssignmentTree = tree;
         // is this a field (and not a local variable)?
         if (TreeUtils.elementFromDeclaration(tree).getKind().isField()) {
-            Set<AnnotationMirror> annotationMirrors =
+            AnnotationMirrorSet annotationMirrors =
                     atypeFactory.getAnnotatedType(tree).getExplicitAnnotations();
             // Fields cannot have commitment annotations.
             for (Class<? extends Annotation> c : atypeFactory.getSupportedTypeQualifiers()) {