+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java
@@ -2998,7 +2998,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
         }
 
         if (ctor.getEnclosingElement().getKind() == ElementKind.ENUM) {
-            Set<AnnotationMirror> enumAnnos = getEnumConstructorQualifiers();
+            AnnotationMirrorSet enumAnnos = getEnumConstructorQualifiers();
             con.getReturnType().replaceAnnotations(enumAnnos);
         }
 
@@ -3018,8 +3018,8 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
      *
      * @return the annotations that should be applied to enum constructors
      */
-    protected Set<AnnotationMirror> getEnumConstructorQualifiers() {
-        return Collections.emptySet();
+    protected AnnotationMirrorSet getEnumConstructorQualifiers() {
+        return new AnnotationMirrorSet();
     }
 
     /**