+++ framework-test/src/main/java/org/checkerframework/framework/test/diagnostics/TestDiagnosticUtils.java
@@ -191,6 +191,49 @@ public class TestDiagnosticUtils {
                 }
             }
         }
+
+        // Check if the message matches detailed message format.
+        // Trim the message to remove leading/trailing whitespace.
+        // Keep separator in sync with SourceChecker.DETAILS_SEPARATOR.
+        String[] diagnosticStrings =
+                Arrays.stream(message.split(" \\$\\$ ")).map(String::trim).toArray(String[]::new);
+        if (diagnosticStrings.length > 1) {
+            // See SourceChecker.detailedMsgTextPrefix.
+            // The parts of the detailed message are:
+
+            // (1) message key;
+            String messageKey = diagnosticStrings[0];
+
+            // (2) number of additional tokens, and those tokens; this depends on the error message,
+            // and an example is the found and expected types;
+            int numAdditionalTokens = Integer.parseInt(diagnosticStrings[1]);
+            int lastAdditionalToken = 2 + numAdditionalTokens;
+            List<String> additionalTokens =
+                    Arrays.asList(diagnosticStrings).subList(2, lastAdditionalToken);
+
+            // (3) the diagnostic position, given by the format (startPosition, endPosition);
+            String pairParens = diagnosticStrings[lastAdditionalToken];
+            // remove the leading and trailing parentheses and spaces
+            String pair = pairParens.substring(2, pairParens.length() - 2);
+            String[] diagPositionString = pair.split(", ");
+            long startPosition = Long.parseLong(diagPositionString[0]);
+            long endPosition = Long.parseLong(diagPositionString[1]);
+
+            // (4) the human-readable diagnostic message.
+            String readableMessage = diagnosticStrings[lastAdditionalToken + 1];
+
+            return new DetailedTestDiagnostic(
+                    file,
+                    lineNo,
+                    kind,
+                    messageKey,
+                    additionalTokens,
+                    startPosition,
+                    endPosition,
+                    readableMessage,
+                    isFixable);
+        }
+
         return new TestDiagnostic(file, lineNo, kind, message, isFixable);
     }