+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@ Version 3.42.0-eisop4 (April ?, 2024)
 
 **Implementation details:**
 
+New method `GenericAnnotatedTypeFactory#addComputedTypeAnnotationsWithoutFlow(Tree, AnnotatedTypeMirror)` that sets
+`useFlow` to `false` before calling `addComputedTypeAnnotations`. Subclasses should override method
+`GenericAnnotatedTypeFactory#addComputedTypeAnnotations(Tree, AnnotatedTypeMirror)` instead.
+Deprecated the `GenericAnnotatedTypeFactory#addComputedTypeAnnotations(Tree, AnnotatedTypeMirror, boolean)` overload.
+
 Improvements in `framework-test` to more consistently handle tests that do not use
 `-Anomsgtext`.